backend/requirements.txt
# =============================================
# MENTAL HEALTH PLATFORM - FULL REQUIREMENTS
# Supports: Backend API, Streamlit UI, Async Indexers (Baltimore, CDC)
# Python 3.10+
# =============================================

# ===== CORE BACKEND / ASYNC FRAMEWORK =====
fastapi==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.35
asyncpg>=0.29.0
psycopg[binary,pool]>=3.2.3
pgvector==0.2.4
alembic>=1.13.0
httpx>=0.27.2
aiohttp>=3.9.1
tenacity==9.0.0
backoff>=2.2.1

# Core async framework
aiofiles>=23.2.1
aiolimiter>=1.1.0 

# ===== CONFIGURATION & ENVIRONMENT =====
pydantic==2.9.2
pydantic-settings>=2.1.0
python-dotenv==1.0.1

# ===== DATA PROCESSING =====
pandas==2.2.2
numpy>=1.26,<2.0
pyarrow==16.1.0
scipy>=1.12.0
scikit-learn>=1.3.0

# ===== STORAGE & CLOUD =====
boto3==1.35.36
# botocore is implied by boto3

# ===== MACHINE LEARNING / LLM INTEGRATION =====
openai>=1.52.0

# ===== VISUALIZATION / STREAMLIT UI =====
streamlit>=1.29.0
matplotlib>=3.8.2
plotly>=5.18.0
seaborn>=0.13.0
altair>=5.0.0
pydeck>=0.8.1b0

# ===== DATA EXPORT / IMPORT =====
openpyxl>=3.1.2
xlsxwriter>=3.1.9
xlrd>=2.0.1
fastparquet>=2023.10.0

# ===== LOGGING / MONITORING =====
structlog>=23.2.0
python-json-logger>=2.0.7
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# ===== UTILITIES / TESTING / DEV TOOLS =====
jsonschema>=4.19.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.27.0
pytest-mock>=3.12.0
black>=23.12.0
ruff>=0.1.8
flake8>=6.0.0
mypy>=1.7.1
types-requests>=2.31.0
types-python-dateutil>=2.8.19
ipython>=8.0.0

# For PubMed
biopython
sodapy
psycopg2-binary

sentence-transformers
scikit-learn
pyarrow

# =============================================
# END OF FILE
# =============================================
